Add files via upload
A console based C++ program to solve the well-known problem "Knight Moves". To be brief, the program performs a variation of the BFS algorithm on a chessboard and finds the smallest possible number of moves a knight should take in order to reach a specified square from its current position.